Renamed staff management component
